refactor(cliente): renombrar comun-tipos como tipos-modelo
Usar un nombre más específico, común tipos es muy genérico
